✨Feat: 채팅방 목록 UI 레이아웃 추가
