+++ src/App.js
@@ -4,6 +4,7 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import User from "./pages/user";
 import Board from "./pages/board";
 import Chat from "./pages/chat";
+import ChatRoom from "./pages/chat/chatroom";
 import Map from "./pages/map";
 import MyPage from "./pages/mypage";
 import LoginPage from "./pages/user";
@@ -28,6 +29,7 @@ function App() {
           <Route path="/user/:uuid" element={<MyPage />} />
           <Route path="/board" element={<Board />} />
           <Route path="/chat" element={<Chat />} />
+          <Route path="/chat/:rood_id" element={<ChatRoom />} />
           <Route path="/map" element={<Map />} />
           <Route path="/mypage" element={<MyPage />} />
         </Routes>
