+++ src/pages/chat/chatitem.css
@@ -1,0 +1,33 @@
+.chatting {
+  display: flex;
+  flex-direction: row;
+}
+
+.chat-thumbnail {
+  width: 48px;
+  height: 48px;
+  border-radius: 50%;
+  object-fit: cover;
+  margin-right: 12px;
+}
+
+.chat-texts {
+  display: flex;
+  flex-direction: column;
+  margin-left: 10px;
+}
+
+.chat-nickname {
+  font-weight: bold;
+  color: #333;
+  margin-bottom: 4px;
+}
+
+.chat-last-message {
+  font-size: 14px;
+  color: #777;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  max-width: 220px;
+}
